+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Allow adding maps to image collection tilesets (#3447)
 * Added file system actions to the tile context menu (#3448)
+* Fixed ability to reset nested string and file properties (#3409)
 
 ### Tiled 1.9.1 (11 August 2022)